fix(ai-features): split capability helpers into client-safe module
The previous commit (ed6b729) added `inferCapabilities` /
`effectiveCapabilities` to `lib/domain/ai/features/router.ts` and
then imported them into the AIFeatureRoutingPage client component.
router.ts has `import "server-only"` and a transitive next/headers
dep via the logger — importing it client-side broke the Turbopack
build on /settings/ai with "next/headers only works in a Server
Component".

Splits the pure helpers into a new dependency-free module
(`./capabilities.ts`) with no imports of its own, so any client
component can safely use them. router.ts re-exports the helpers
under the same names for existing server callers, then imports
them locally for `modelSatisfiesCapabilities` /
`listCompatibleModels`.

tsc doesn't enforce the server/client boundary — that's the bundler's
job — so this slipped past the local typecheck. Adding the dev-server
boot to the pre-push verification checklist for any change that
touches a `"server-only"` module.
