+++ lib/domain/ai/features/router.ts
@@ -106,58 +106,11 @@ export async function resolvePrimaryRoute(
   return all[0] ?? null;
 }
 
-/**
- * Infer capability flags from a model's id when the saved
- * `capabilities` array is missing or incomplete.
- *
- * Safety net for older Connection entries that were saved before the
- * fetcher's catalog augmentation existed (or for entries the user
- * added manually via the "Model ID" input, which has no capability
- * field). Lets feature routing find them as compatible pairs without
- * forcing the user to re-add their models.
- *
- * Only adds capabilities — never removes — so explicit `capabilities`
- * on the model still win. Patterns are deliberately conservative; we
- * recognize widely-documented model id stems rather than every
- * variant.
- */
-export function inferCapabilities(modelId: string): string[] {
-  // Strip a namespace prefix so `openai/dall-e-3` matches the same
-  // patterns as `dall-e-3`.
-  const slash = modelId.indexOf("/");
-  const bare = slash >= 0 ? modelId.slice(slash + 1) : modelId;
-  const out: string[] = [];
-
-  // Image generation models. These ids are stable across providers and
-  // gateways — DALL·E + GPT Image (OpenAI), Imagen (Google), FLUX
-  // (fal.ai / Together / Fireworks).
-  if (
-    /^dall-e/i.test(bare) ||
-    /^gpt-image/i.test(bare) ||
-    /^imagen/i.test(bare) ||
-    /\bflux\b/i.test(bare) ||
-    /^stable-diffusion/i.test(bare) ||
-    /^sdxl/i.test(bare)
-  ) {
-    out.push("image-generation");
-  }
-
-  return out;
-}
-
-/**
- * Effective capability set for a saved model: union of the explicit
- * `capabilities` array (catalog/fetcher-derived) and any flags
- * inferred from the model id. Use this anywhere a feature-routing
- * filter wants to know what a model can do.
- */
-export function effectiveCapabilities(
-  model: { id: string; capabilities?: string[] },
-): Set<string> {
-  const have = new Set<string>(model.capabilities ?? []);
-  for (const inferred of inferCapabilities(model.id)) have.add(inferred);
-  return have;
-}
+// Capability helpers live in `./capabilities.ts` so client components
+// can import them without dragging server-only deps through this file.
+// Re-export here so existing server-side callers keep working.
+export { inferCapabilities, effectiveCapabilities } from "./capabilities";
+import { effectiveCapabilities } from "./capabilities";
 
 function modelSatisfiesCapabilities(
   connection: ConnectionView,
